+++ lib/airbrake/rails/http_client.rb
@@ -3,7 +3,7 @@ class HTTPClient
   alias do_get_without_airbrake do_get_block
 
   def do_get_block(request, proxy, connection, &block)
-    Airbrake::Rack.capture_http_performance do
+    Airbrake::Rack.capture_timing(:http) do
       do_get_without_airbrake(request, proxy, connection, &block)
     end
   end