+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Airbrake Changelog
   ([#960](https://github.com/airbrake/airbrake/pull/960))
 * Rails: added support for Typhoeus for HTTP performance breakdown
   ([#961](https://github.com/airbrake/airbrake/pull/961))
+* Added `Airbrake::Rack.capture_timing` and `Airbrake::Rack::Instrumentable` for
+  manual performance measurements
+  ([#965](https://github.com/airbrake/airbrake/pull/965))
 
 ### [v9.1.0][v9.1.0] (April 17, 2019)